+++ config/auth.php
@@ -67,13 +67,8 @@ return [
     'providers' => [
         'users' => [
             'driver' => 'eloquent',
-            'model' => App\User::class,
+            'model' => App\Models\User::class,
         ],
-
-        // 'users' => [
-        //     'driver' => 'database',
-        //     'table' => 'users',
-        // ],
     ],
 
     /*
@@ -98,7 +93,5 @@ return [
             'expire' => 60,
         ],
     ],
-
-    'cookie_key' => env('COOKIE_KEY'),
 
 ];