+++ legacy/logout.php
@@ -7,6 +7,7 @@ include($ugamela_root_path . 'common.php');
 
 includeLang('logout');
 
-setcookie($game_config['COOKIE_NAME'], "", time()-100000, "/", "", 0);
+Auth::logout();
+Session::save();
 
 message($lang['see_you'], $lang['session_closed'], "login", 2);