+++ legacy/options.php
@@ -136,11 +136,12 @@ if ($_POST && $mode == "change") {
     `deltime` = '$Del_Time'
     WHERE `id` = '$iduser' LIMIT 1", "users");
 
-    if (isset($_POST["db_password"]) && md5($_POST["db_password"]) == $user["password"]) {
+    if (isset($_POST["db_password"]) && Hash::check($_POST["db_password"], $user["password"])) {
         if ($_POST["newpass1"] == $_POST["newpass2"] && strlen($_POST["newpass1"]) >= 4) {
-            $newpass = md5($_POST["newpass1"]);
+            $newpass = bcrypt($_POST["newpass1"]);
             doquery("UPDATE {{table}} SET `password` = '{$newpass}' WHERE `id` = '{$user['id']}' LIMIT 1", "users");
-            setcookie(COOKIE_NAME, "", time()-100000, "/", "", 0); //le da el expire
+            Auth::logout();
+            Session::save();
             message($lang['succeful_changepass'], $lang['changue_pass']);
         }
     }
@@ -148,7 +149,8 @@ if ($_POST && $mode == "change") {
         $query = doquery("SELECT id FROM {{table}} WHERE username='{$_POST["db_character"]}'", 'users', true);
         if (!$query) {
             doquery("UPDATE {{table}} SET username='{$username}' WHERE id='{$user['id']}' LIMIT 1", "users");
-            setcookie(COOKIE_NAME, "", time()-100000, "/", "", 0); //le da el expire
+            Auth::logout();
+            Session::save();
             message($lang['succeful_changename'], $lang['changue_name']);
         }
     }